ref: Move SenderLimitTest to ConferenceTest.
